+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Nomenclature :
 
 ### Unreleased
 
+- Updated Observable on delete fonctions (tenant, owner, property)
+
 ### [0.1.15] - 2023-04-26
 
 #### Updated